requirements.txt
certifi==2021.10.8
minio==7.1.3
python-dotenv==0.19.2
urllib3==1.26.8
